+++ tests/caliban/test_util.py
@@ -89,6 +89,79 @@ class UtilTestSuite(unittest.TestCase):
     }]
     self.assertListEqual(result, expected)
 
+  def test_compound_key_handling(self):
+    """ tests the full assembly line transforming a configuration dictionary
+    including compound keys into a list of dictionaries for passing to the script
+    """
+
+    tests = [
+        {
+          'input': {'[a,b]': [['c', 'd'], ['e', 'f']]},
+          'after_tupleization': {('a', 'b'): [('c', 'd'), ('e', 'f')]},
+          'after_dictproduct': [{('a', 'b'): ('c', 'd')},
+                                {('a', 'b'): ('e', 'f')}],
+          'after_expansion': [{'a':'c', 'b':'d'},
+                                      {'a':'e', 'b':'f'}]
+        },
+        {
+          'input': {'[a,b]': ['c', 'd']},
+          'after_tupleization': {('a','b'): ('c','d')},
+          'after_dictproduct': [{('a','b'): ('c','d')}],
+          'after_expansion': [{'a': 'c', 'b': 'd'}]
+        },
+        {
+          'input': {'hi': 'there', '[k1,k2]': [['v1a', 'v2a'], ['v1b', 'v2b']]},
+          'after_tupleization': {'hi': 'there', ('k1', 'k2'): [('v1a', 'v2a'), ('v1b', 'v2b')]},
+          'after_dictproduct': [{'hi': 'there', ('k1', 'k2'): ('v1a', 'v2a')},
+                                      {'hi': 'there', ('k1', 'k2'): ('v1b', 'v2b')}],
+          'after_expansion': [{'hi': 'there', 'k1': 'v1a', 'k2': 'v2a'},
+                                      {'hi': 'there', 'k1': 'v1b', 'k2': 'v2b'}]
+        },
+        {
+          'input': {'hi': 'there', '[a,b]': ['c', 'd']},
+          'after_tupleization': {'hi': 'there', ('a', 'b'): ('c', 'd')},
+          'after_dictproduct': [{'hi': 'there', ('a', 'b'): ('c', 'd')}],
+          'after_expansion': [{'hi': 'there', 'a': 'c', 'b': 'd'}]
+        },
+        {
+          'input': {'[a,b]': [0, 1]},
+          'after_tupleization': {('a', 'b'): (0, 1)},
+          'after_dictproduct': [{('a', 'b'): (0, 1)}],
+          'after_expansion': [{'a': 0, 'b': 1}]
+        },
+        {
+          'input': {'[a,b]': [[0, 1]]},
+          'after_tupleization': {('a', 'b'): [(0, 1)]},
+          'after_dictproduct': [{('a', 'b'): (0, 1)}],
+          'after_expansion': [{'a': 0, 'b': 1}]
+        },
+        {
+          'input': {'hi': 'blueshift', '[a,b]': [[0, 1]]},
+          'after_tupleization': {'hi': 'blueshift', ('a', 'b'): [(0, 1)]},
+          'after_dictproduct': [{'hi': 'blueshift', ('a', 'b'): (0,1)}],
+          'after_expansion': [{'hi': 'blueshift', 'a': 0, 'b': 1}]
+        }
+            ]
+
+    def check_tupleization(test_dict):
+      self.assertDictEqual(test_dict['after_tupleization'],
+                           u.tupleize_dict(test_dict['input']))
+
+
+    def check_dictproduct(test_dict):
+      self.assertListEqual(test_dict['after_dictproduct'],
+                           list(u.dict_product(test_dict['after_tupleization'])))
+
+    def check_expansion(test_dict):
+      self.assertListEqual(test_dict['after_expansion'],
+                           list(u.expand_compound_dict(test_dict['after_dictproduct'])))
+
+    for test in tests:
+      check_tupleization(test)
+      check_dictproduct(test)
+      check_expansion(test)
+
+
   @given(st.integers())
   def test_compose(self, x):
     """Functions should compose; the composed function accepts any arguments that the rightmost function accepts."""
@@ -288,7 +361,7 @@ class UtilTestSuite(unittest.TestCase):
   def test_key_value_label(self):
     """unit tests for specific cases of key and value label conversion."""
     self.assertEqual("face", u.key_label("--face"))
-    self.assertEqual("face", u.key_label("------fA.!! ce"))
+    self.assertEqual("fa_ce", u.key_label("------fA.!! ce"))
 
     # Empty string roundtrips.
     self.assertEqual("", u.key_label(""))